final update

+++ src/Controller/AdminController.php
@@ -65,19 +65,32 @@ class AdminController
         }
     }
 
+    public static function deleteUserVerif()
+    {
+        if (self::isAdmin()) {
+            $token = DefaultController::generer_token('delete_user');
+            echo $token;
+        }
+    }
+
     public static function deleteUser()
     {
         if (self::isAdmin()) {
-            if (isset($_POST["id"])) {
-                $base = Repository::connect();
-                $userRepository = new UserRepository($base);
-                $user = $userRepository->delete($_POST["id"]);
+            if (DefaultController::verifier_token(120, 'delete_user')) {
+                if (isset($_POST["id"])) {
+                    $base = Repository::connect();
+                    $userRepository = new UserRepository($base);
+                    $user = $userRepository->delete($_POST["id"]);
 
-                if ($user) {
-                    return true;
-                } else {
-                    return false;
+                    if ($user) {
+                        return true;
+                    } else {
+                        return false;
+                    }
                 }
+            } else {
+                //envoi d'un message
+                DefaultController::alertMessage("danger", "Mauvais Token");
             }
         }
     }
@@ -101,16 +114,16 @@ class AdminController
     {
         if (self::isAdmin()) {
             if (DefaultController::verifier_token(120, 'modifier_user')) {
-            if (isset($_POST["id"]) && isset($_POST["email"]) && isset($_POST["admin"])) {
-                $base = Repository::connect();
-                $userRepository = new UserRepository($base);
-                if (isset($_POST['mdp'])) {
-                    $userRepository->modifyByIdWithMdp($_POST["id"], $_POST["email"], $_POST["admin"], password_hash($_POST["mdp"], PASSWORD_ARGON2I));
-                } else {
-                    $userRepository->modifyById($_POST["id"], $_POST["email"], $_POST["admin"]);
+                if (isset($_POST["id"]) && isset($_POST["email"]) && isset($_POST["admin"])) {
+                    $base = Repository::connect();
+                    $userRepository = new UserRepository($base);
+                    if (isset($_POST['mdp'])) {
+                        $userRepository->modifyByIdWithMdp($_POST["id"], $_POST["email"], $_POST["admin"], password_hash($_POST["mdp"], PASSWORD_ARGON2I));
+                    } else {
+                        $userRepository->modifyById($_POST["id"], $_POST["email"], $_POST["admin"]);
+                    }
+                    return $userRepository;
                 }
-                return $userRepository;
-            }
             } else {
                 //envoi d'un message
                 DefaultController::alertMessage("danger", "Mauvais Token");
@@ -130,12 +143,12 @@ class AdminController
     {
         if (self::isAdmin()) {
             if (DefaultController::verifier_token(120, 'ajouter_user')) {
-            if (isset($_POST["email"]) && isset($_POST["mdp"]) && isset($_POST["admin"])) {
-                $base = Repository::connect();
-                $userRepository = new UserRepository($base);
-                $userRepository->save($_POST["email"], password_hash($_POST["mdp"], PASSWORD_ARGON2I), $_POST["admin"]);
-                return $userRepository;
-            }
+                if (isset($_POST["email"]) && isset($_POST["mdp"]) && isset($_POST["admin"])) {
+                    $base = Repository::connect();
+                    $userRepository = new UserRepository($base);
+                    $userRepository->save($_POST["email"], password_hash($_POST["mdp"], PASSWORD_ARGON2I), $_POST["admin"]);
+                    return $userRepository;
+                }
             } else {
                 //envoi d'un message
                 DefaultController::alertMessage("danger", "Mauvais Token");
@@ -169,6 +182,7 @@ class AdminController
             require __DIR__ . '/../View/admin/afficherDispo.php';
         }
     }
+
     public static function deleteDispoVerif()
     {
         if (self::isAdmin()) {
@@ -176,21 +190,22 @@ class AdminController
             echo $token;
         }
     }
+
     public static function deleteDispo()
     {
         if (self::isAdmin()) {
             if (DefaultController::verifier_token(120, 'delete_dispo')) {
-            if (isset($_POST["idSalle"]) && isset($_POST["idCreneau"])) {
-                $base = Repository::connect();
-                $dispoRepository = new DispoRepository($base);
-                $user = $dispoRepository->deleteByArguments($_POST["idSalle"], $_POST["idCreneau"]);
+                if (isset($_POST["idSalle"]) && isset($_POST["idCreneau"])) {
+                    $base = Repository::connect();
+                    $dispoRepository = new DispoRepository($base);
+                    $user = $dispoRepository->deleteByArguments($_POST["idSalle"], $_POST["idCreneau"]);
 
-                if ($user) {
-                    return true;
-                } else {
-                    return false;
+                    if ($user) {
+                        return true;
+                    } else {
+                        return false;
+                    }
                 }
-            }
             } else {
                 //envoi d'un message
                 DefaultController::alertMessage("danger", "Mauvais Token");
@@ -221,14 +236,14 @@ class AdminController
     {
         if (self::isAdmin()) {
             if (DefaultController::verifier_token(120, 'ajouter_dispo')) {
-            if (isset($_POST["jour"]) && isset($_POST["idSalle"]) && isset($_POST["idCreneau"])) {
-                $base = Repository::connect();
-                $dispoRepository = new DispoRepository($base);
-                $dispoDate = explode('/', $_POST["jour"]);
-                $dispoDateString = $dispoDate[0] . "-" . $dispoDate[1] . "-" . $dispoDate[2];
-                $dispoRepository->add($dispoDateString, $_POST["idSalle"], $_POST["idCreneau"]);
-                return $dispoRepository;
-            }
+                if (isset($_POST["jour"]) && isset($_POST["idSalle"]) && isset($_POST["idCreneau"])) {
+                    $base = Repository::connect();
+                    $dispoRepository = new DispoRepository($base);
+                    $dispoDate = explode('/', $_POST["jour"]);
+                    $dispoDateString = $dispoDate[0] . "-" . $dispoDate[1] . "-" . $dispoDate[2];
+                    $dispoRepository->add($dispoDateString, $_POST["idSalle"], $_POST["idCreneau"]);
+                    return $dispoRepository;
+                }
             } else {
                 //envoi d'un message
                 DefaultController::alertMessage("danger", "Mauvais Token");
@@ -261,12 +276,12 @@ class AdminController
     {
         if (self::isAdmin()) {
             if (DefaultController::verifier_token(120, 'ajouter_salle')) {
-            if (isset($_POST["numSalle"]) && isset($_POST["nbPlace"]) && isset($_POST["dispo"])) {
-                $base = Repository::connect();
-                $salleRepository = new SalleRepository($base);
-                $salleRepository->save($_POST["numSalle"], $_POST["nbPlace"], $_POST["dispo"]);
-                return 0;
-            }
+                if (isset($_POST["numSalle"]) && isset($_POST["nbPlace"]) && isset($_POST["dispo"])) {
+                    $base = Repository::connect();
+                    $salleRepository = new SalleRepository($base);
+                    $salleRepository->save($_POST["numSalle"], $_POST["nbPlace"], $_POST["dispo"]);
+                    return 0;
+                }
             } else {
                 //envoi d'un message
                 DefaultController::alertMessage("danger", "Mauvais Token");
@@ -281,21 +296,22 @@ class AdminController
             echo $token;
         }
     }
+
     public static function deleteSalle()
     {
         if (self::isAdmin()) {
             if (DefaultController::verifier_token(120, 'delete_salle')) {
-            if (isset($_POST["id"])) {
-                $base = Repository::connect();
-                $salleRepository = new SalleRepository($base);
-                $salle = $salleRepository->delete($_POST["id"]);
+                if (isset($_POST["id"])) {
+                    $base = Repository::connect();
+                    $salleRepository = new SalleRepository($base);
+                    $salle = $salleRepository->delete($_POST["id"]);
 
-                if ($salle) {
-                    return true;
-                } else {
-                    return false;
+                    if ($salle) {
+                        return true;
+                    } else {
+                        return false;
+                    }
                 }
-            }
             } else {
                 //envoi d'un message
                 DefaultController::alertMessage("danger", "Mauvais Token");
